Prefer _priv.h over _pub.h
