+++ src/func/gig_callback.c
@@ -19,8 +19,8 @@
 #include <ffi.h>
 #include "../core.h"
 #include "gig_argument.h"
-#include "gig_callback.h"
-#include "gig_function.h"
+#include "gig_callback_priv.h"
+#include "gig_function_priv.h"
 
 typedef struct _GigCallback GigCallback;
 struct _GigCallback
