+++ src/func/gig_function_priv.h
@@ -13,15 +13,21 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-#ifndef GIG_FUNCTION_PUB_H
-#define GIG_FUNCTION_PUB_H
+#ifndef GIG_FUNCTION_H
+#define GIG_FUNCTION_H
 
 #include <girepository.h>
 #include <libguile.h>
+#include "gig_arg_map.h"
+#include "gig_function.h"
 
-GType *gig_function_get_arg_gtypes(GICallableInfo *info, size_t *len);
-SCM gig_function_define(GType type, GICallableInfo *info, const char *_namespace);
+typedef SCM (*GigGsubr)(void);
 
-void gig_init_function(void);
+extern SCM ensure_generic_proc;
+extern SCM top_type;
+extern SCM sym_self;
+
+SCM gig_callable_invoke(GICallableInfo *callable_info, void *callable, GigArgMap *amap,
+                        const char *name, GObject *self, SCM args, GError **error);
 
 #endif
