+++ src/type.h
@@ -1,8 +1,8 @@
 #ifndef GI_TYPE_H
 #define GI_TYPE_H
-#include "type/gig_constant_pub.h"
-#include "type/gig_flag_pub.h"
-#include "type/gig_object_pub.h"
-#include "type/gig_signal_pub.h"
-#include "type/gig_type_pub.h"
+#include "type/gig_constant.h"
+#include "type/gig_flag.h"
+#include "type/gig_object.h"
+#include "type/gig_signal.h"
+#include "type/gig_type.h"
 #endif
