+++ src/type/gig_boxed.c
@@ -15,7 +15,7 @@
 
 #include <assert.h>
 #include "../core.h"
-#include "gig_type_private.h"
+#include "gig_boxed.h"
 
 static slist_t *_boxed_funcs = NULL;
