+++ src/type/gig_closure.c
@@ -15,7 +15,7 @@
 
 #include "gig_closure.h"
 #include "gig_value.h"
-#include "gig_type.h"
+#include "gig_type_priv.h"
 
 typedef struct _GigClosure GigClosure;
