+++ src/type/gig_constant_priv.h
@@ -1,4 +1,4 @@
-// Copyright (C) 2022 Michael L. Gran
+// Copyright (C) 2018, 2022 Michael L. Gran
 
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -13,11 +13,9 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-#ifndef GIG_CALLBACK_PUB_H
-#define GIG_CALLBACK_PUB_H
+#ifndef GIG_CONSTANT_PRIV_H
+#define GIG_CONSTANT_PRIV_H
 
-#include "../core.h"
-
-GIG_API void gig_init_callback(void);
+#include "gig_constant.h"
 
 #endif
