+++ src/type/gig_flag.c
@@ -16,8 +16,8 @@
 #include <inttypes.h>
 #include <libguile.h>
 #include "../core.h"
-#include "gig_flag.h"
-#include "gig_type.h"
+#include "gig_flag_priv.h"
+#include "gig_type_priv.h"
 
 static SCM enum_to_number;
 static SCM flags_to_number;
