+++ src/type/gig_object.h
@@ -16,6 +16,11 @@
 #ifndef GIG_OBJECT_H
 #define GIG_OBJECT_H
 
-#include "gig_object_pub.h"
+#include <libguile.h>
+#include <girepository.h>
+#include "../core.h"
+
+SCM gig_property_define(GType type, GIPropertyInfo *info, const char *namespace);
+GIG_API void gig_init_object(void);
 
 #endif
