+++ src/type/gig_type.h
@@ -19,24 +19,28 @@
 #include <stdbool.h>
 #include <girepository.h>
 #include "../core.h"
-#include "gig_type_pub.h"
 
-SCM gig_enum_type(void);
-SCM gig_flags_type(void);
-SCM gig_object_type(void);
-SCM gig_paramspec_type(void);
-SCM gig_value_type(void);
-SCM gig_closure_type(void);
+GType gig_type_get_c_array_type(void);
 
-char *gig_type_class_name_from_gtype(GType gtype);
+#define G_TYPE_PRIV_C_ARRAY (gig_type_get_c_array_type())
 
-void gig_type_register(GType gtype, SCM stype);
-SCM gig_type_define_full(GType gtype, SCM extra_supers);
+typedef void *(*GigTypeRefFunction)(void *);
+typedef void (*GigTypeUnrefFunction)(void *);
 
-GType gig_type_get_gtype_from_obj(SCM x);
-SCM gig_type_get_scheme_type_with_info(GIRegisteredTypeInfo *info);
+bool gig_type_is_registered(GType gtype);
 
-bool gig_type_check_object(SCM obj);
-void *gig_type_peek_typed_object(SCM obj, SCM expected);
+SCM gig_type_define(GType gtype);
+void gig_type_define_fundamental(GType type, SCM extra_supers,
+                                 GigTypeRefFunction ref, GigTypeUnrefFunction unref);
+SCM gig_type_define_with_info(GIRegisteredTypeInfo *info, SCM slots);
+SCM gig_type_get_scheme_type(GType gtype);
+void *gig_type_peek_object(SCM obj);
+bool gig_type_check_typed_object(SCM obj, SCM expected_type);
+SCM gig_type_transfer_object(GType gtype, void *obj, GITransfer transfer);
+SCM scm_from_gtype(GType x);
+GType scm_to_gtype(SCM x);
+GType scm_to_gtype_full(SCM x, const char *subr, int argpos);
+
+GIG_API void gig_init_types(void);
 
 #endif
